Refactored the email auth functions and seperated them and. Fix : Anonymous account deletion after new account.
